+++ src/main/java/org/folio/sidecar/integration/keycloak/filter/KeycloakAuthorizationFilter.java
@@ -14,10 +14,10 @@ import static org.folio.sidecar.utils.RoutingUtils.getParsedSystemToken;
 import static org.folio.sidecar.utils.RoutingUtils.getParsedToken;
 import static org.folio.sidecar.utils.RoutingUtils.getScRoutingEntry;
 import static org.folio.sidecar.utils.RoutingUtils.getTenant;
-import static org.folio.sidecar.utils.RoutingUtils.hasNoPermissionsRequired;
 import static org.folio.sidecar.utils.RoutingUtils.isSelfRequest;
 import static org.folio.sidecar.utils.RoutingUtils.isSystemRequest;
 import static org.folio.sidecar.utils.RoutingUtils.isTimerRequest;
+import static org.folio.sidecar.utils.RoutingUtils.requiresNoNamedPermission;
 
 import com.github.benmanes.caffeine.cache.Cache;
 import io.quarkus.security.ForbiddenException;
@@ -77,7 +77,7 @@ public class KeycloakAuthorizationFilter implements IngressRequestFilter, CacheI
 
   @Override
   public boolean shouldSkip(RoutingContext rc) {
-    return !isTimerRequest(rc) && (isSystemRequest(rc) || hasNoPermissionsRequired(rc)) || isSelfRequest(rc);
+    return !isTimerRequest(rc) && (isSystemRequest(rc) || requiresNoNamedPermission(rc)) || isSelfRequest(rc);
   }
 
   @Override